Move activities types where they belong
