+++ packages/ui/src/proposals/hooks/useProposalsActivities.ts
@@ -1,5 +1,5 @@
-import { asProposalActivities } from '../components/ProposalsActivities'
 import { useGetProposalsEventsQuery } from '../queries/__generated__/proposalsEvents.generated'
+import { asProposalActivities } from '../types/ProposalsActivities'
 
 export const useProposalsActivities = () => {
   const { data, loading } = useGetProposalsEventsQuery()
